Makefile, set MCU at top and remove -nostdlib
Define MCU in one place at the top of the file.  Also allow it to be
overridden by the command line to compile for a different MCU without
changing the Makefile.
Reduce confusion, remove -nostdlib as it isn't obvious that it is
only being passed when compiling (which isn't linking so doesn't use
it), and isn't being passed to the linker (where it would fail if it
was, because routines from the library are used).