Un-inlined and weakened the definitions of utoa and itoa.
The problem with the definitions of utoa and itoa is that they
are implemented in the latest version newlib. This means that
these functions will be redefined from a linker point of view,
if the linker uses versions of newlib, in which utoa and itoa are
implemented.

As the weak attribute is not compatible with static and inline,
both static and inline was removed from the definition.

The implementation of utoa and itoa was moved to nonstd.c.