T3.5/T3.6 Beta: Optional Allow SDCard pins
With the new beta boards, for those users who don't need an SDCard for
their project, it would be great to allow them to use the 6 IO pins
associated with the SDCARD for other purposes.  This would probably
require a cheap adapter, but these are nice signals to be able to be
used.  They include:

SPI1: SCK, MISO, MOSI, CS0-2
Serail2: TX/RX
Serial4: TX/RX
I2C1: SCL/SDA

This change adds support for GPIO, SPI, Serial2 and Serial4