T3.6 - If we are going to use HSRUN, read serial number before
If we are runnig at over 120mhz and turn on HSRUN mode, we can not read
in the USB serial number.

This is one attempt to see if we can read it in before we switch to
HSRUN mode.

My quick couple of test appeared to work, but may be cleaner ways