Transfercount plus TNI's aligned attribute
I inclded TNI's aligned attribute fix as it is needed for DMA to work at
all with T-LC...

The new fix with this delta, is there is a problem with the
transferCount method of the DMABase class for T3.x boards.

In particular it checks to see if the count < 512 and does somethig
different which can screw up.   It assmes that if the count is under 512
you are using the ELINK version of the output, which is fine as long as
your previous calls were also under 512.

But with current code, if You call this function in an order like:
Count 6  - Biter and Citer will be set to 6
Count 512 - Biter and Citer will be set to 512
Count 6 - Biter and Citer will be set to 518

Fix is to instead of check for >= 512 to instead check for the ELINK bit
being set then use the 15 bit version else the 7 bit.

Note: potentially should check in the ELINK case that value is < 512,
else it won't work... But not sure what to do in error case?  Currently
if you call with > 32767 it just returns without doing somehing,

But again function does not return any error status, so app will also
likely hang or at best not transfer...