T4 - SPI avr_emulation - only enough for Adafruit_gfx...
QA quick and dirty import of the AVR emulation stuff for SPI registers.

So far none of SPCR is implemented
in SPSR only getting status of SPIF is setup
SPDR  reads and writes to LPSPI4...

So, it allows for some simple SPI write functions and wait for the write to complete and read the result.

WIth this Adafruit_ILI9341 graphic test runs