Explicitely delcare TMR channel registers in IMXRT_TMR_t to make encapsulation of the registers in a user class easier
I suggest to explicitely delcare a (say)  IMXRT_TMR_CH_t struct which collects the registers for a TMR  channel instead of implicitely defining this struct in the declaration of IMXRT_TMR_t.

This would allow for an easier definition of user classes modelling a TMR channel.
Not a big deal, could also do that mayself the class code but it would be more convenient if this would be done in imxrt.h