T4 = Hardware serial Half duplex support
As an experiment added in half duplex mode into the Hardware Serial class.

If you pass in the SERIAL_HALF_DUPLEX
flag as part of the format on the begin method, it will configure the TX pin to be
half duplex mode.  When you do writes it will turn the pin to TX pin and when Tx is emplty ... will convert back to RX mode