T4 SerialX - Watermark - look at UART
If we are using a RX WATER value > 0
Add code to SerialX.avail, read and peek

for avail add in  count of entries in RX FIFO.

For Read, if our queue is empty, and FIFO is not empty, read first element and return it.

For Peek if queue is empty and FIFO is not empty, read it and reinitialize queue to have this item as first item in queue and return it.