Update RAW HID TX buffer to be in DMAMEM
Decided to go ahead and move the TX Buffer there as well
Did a quick check in earlier test program to echo the contents back.

The other test program was already setup to receive data on the HID and does hex dump of buffer which looks correct.  So far so good.