T4 Yield speedups by knowing if default serialEventX
Found a way that appears to work to detect if a sketch has our default serialEvantX functions, and that the users code has not implemented their own.  Before on SerialX.begin() I would add that serial ports code to table of function calls to do the if SerialX.available() call serialEvernt..

But I now have a hacked up version where each of the event functions are in their own source file, along with a variable that is defined.  Elsewhere I have those same variables defined with an attribute of ((weak)), so if our default implemention is pulled in it is defined with a value of 1, if the weak version is used because the user has their own implemention, then  it is defined as a 0... So I can detect this at the begin method and only add the code to do the checks in yield if the user implemented it.

WARNING this also brought in changes for using XBar pins for CTS and RX pins... As I did not wish to split back out and have you maybe have to manually merge.