Teensy 3 & 4 std=c++17 compatibility
This change allows compiling the Teensy 3 & 4 cores with the
`arm-none-eabi-g++` `-std=c++17` flag.

It brings in the same `itoa.c` and `itoa.h` used in these cores:
- https://github.com/arduino/ArduinoCore-samd/tree/master/cores/arduino
- https://github.com/adafruit/ArduinoCore-samd/tree/master/cores/arduino

Test: Compiled example sketches using Teensyduino 1.53 with these core
changes. No changes to boards.txt or platform.txt which uses
`-std=gnu++14`.

Motivation for this change is to allow building the Teensy cores for use
in other frameworks. For our use case is building Pigweed for Teensys:
https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/18001/