T3.x/LC-SerialX addMemoryFor...
An experiement to see how much work again it would take to add the T4 methods to allow user sketches to increase the size of Serial ports RX and TX buffers, without having to edit the Source code which does it then for all sketches.

T3.x SerialN Add extra memory capabilities for Serial1-6

Also

Fixed serialEventX yield handling when you call begin multiple times

There is an issue that if you are not using the serialEvent type code and you call SerialX.begin multiple times, it will add that item to the process list multiple times, which can overrun the memory and cause crash or the like.

Fix was to check the list first and see if we are already in the list.