Update the T4.x Imxrt.h file for flexio having 8 timers/shifters
As mentioned in a few places, it appears like the T4.x FlexIO objects have 8 timers and shifters, which differs from most of their documentation, in the PARAM register, the document shows default which is READ only as 4, but the actual values are 8. And their description of Parallel Interface P2894 talks about Shifter 4 and 7...

Functionaly this should be no different, as I simply increased the size of each of the arrays from 4 to 8 in the structure and then decremented the unused array after each one of these by 4... So either way if you ask for example for SHIFTCTL[7], it should give you the same memory location, but without this change the compiler would probably issue a warning.