TMOD - Schematics P38 and 39 are swapped compared to T4
Most of the code assumed that P38 and 39 on The Teensy_MICROMOD which is part of the SDIO pins was the same as it was on T4.  As all of the other pins  used on the SDIO are the same.  However on TMOD pin 38 aand 39 are swapped.

So needed to udpdate the Pin table, the RX/TX alternate pins for Serial5, PWM and the like