T4 DMAChannel  - Implement triggerContinuously
Paul,

In the arm_dcache_delete testing thread, I wanted to setup a test example where I wanted to copy memory using DMA into the area to test what areas are preserved or not preserved and to make sure that the different solutions worked as to get the updated memory after the DMA transfer.

So I tried setting up DMA operation like:
```
 dma.sourceBuffer((uint8_t*)low_buffer, sizeof(low_buffer));
  dma.destinationBuffer((uint8_t*)dmam.dmaBuffer, sizeof(dmam.dmaBuffer));
  dma.transferCount(sizeof(dmam.dmaBuffer));
  dma.disableOnCompletion();
  dma.enable();
  dma.clearComplete();
  dma.triggerContinuously();
```
And only 1 byte transferred.  I loooked at the triggerContinuously implementation and found it was a TODO and was doing nothing.

So i did more or less your comments and in some cases it appear to work.

However it only appears to work when the destination buffer is 32 byte aligned.

But I don't think that is specific to this change.   Decided to go ahead and issue this PR in case someone else tries this.