Prevent update owner from being moved off update list if it's disconnected
Lots of debug code still present here. It appears that once destruction is initiated an update interrupt can prove fatal, even though it would be fine once our destructor code has run. It looks as if the compiler may have made some assumptions / modifications prior to calling our destructor which break the system. The solution appears to be to either disable the interrupt or clear the active flag as early as possible - within ~AudioStream() is too late.