T4. usb_mtp.c - on RX look at actual count of bytes transferred and use it
We were running into a case where the host would send us a 0 length packet if on a SendObject the file fit exactly in full packets... Then it would send one that said no more data on that file.

Our code did not look in the transfer->status field to extract the count transferred.
The code now extracts that and only copies that many bytes and returns it to caller.