MTP Teensy4 - MTP.loop() would wait about 1ms with no messages
We passed in -0 for TO, but the code was waiting for systeick_miillis_count to change before we would return.  This was very noticeable with test sketches.

Also impacting some other sketches as well.  May also need to update for T3.x but I think that is contained within MTP_Teensy