Implement finer-grained control of event responses
Use triggerEventNextYield() to ensure the event will be responded to on the next yield().

Works by dividing the triggered event list into two sections; at the next yield() call, firstYield to nextYield will all be actioned, plus the first of nextYield to lastYield.

Also adds EventResponder::isActive() to determine if any EventTypeYield functions have been attached. This may be of use if users are finding yield() is slower than expected.

keywords.txt needs updating, but isn't in the scope of cores.