Serial port half duplex if inverted should Pull down not up
As per a few threads up on the forum, such as:
https://forum.pjrc.com/threads/72740-Serial-Half-Duplex-Support

I believe this will also hit, it trying to convert DUFF library for SDI-12 to use Serial code in core versus roll your own.

I did quick and dirty test sketch to try it on different uarts.
Only tried on Serial1 on T4.x as same code for all uarts.

Tried on all of them on T3.5 except 6, as was lazy and did not want to get to the bottom pad.  Dif confirm it compiles for T3.6, T3.2 and LC

```
#define SERIALX Serial1
void setup() {
  // put your setup code here, to run once:
  //SERIALX.begin(1200, SERIAL_7E1 | SERIAL_HALF_DUPLEX);
  SERIALX.begin(1200, SERIAL_7E1_RXINV_TXINV | SERIAL_HALF_DUPLEX);
  pinMode(LED_BUILTIN, OUTPUT);
}
int loop_count = 0;
void loop() {
  SERIALX.printf("Loop count: %u\n", ++loop_count);
  digitalToggleFast(LED_BUILTIN);
  delay(1000);
}
```