Create an audio block pre-filled with silence
In theory passing a NULL pointer to an audio block implies a silent block, i.e. all zeroes. In many cases this can be dealt with simply, but on other occasions code is significantly simplified if a genuine audio block pre-filled with silent data is available. An example is a reverb tail when its source suddenly starts emitting NULL blocks; there are also multiple instances where hardware output objects have private zerodata[] arrays which could be replaced by using this common structure.

This modification adds an AudioStream::allocate_silent() function, which yields a "genuine" silent block for use on such occasions. It has the restriction that its data are read-only, but it can be transmitted and released as with any other block, so no special treatment is needed by "downstream" code. Any number of copies may be allocated at one time (the reference count is never modified), even if the block pool is exhausted.

At the time of writing this extension is not used within the (separately maintained) Audio library, but once part of the cores it will be possible to update the Audio library to make use of it, hopefully improving its clarity and efficiency.