Fix T3.x SerialEventX when format is passed in.
Before there were two different begin Methods that were wrappers
to the c begin and format functions.  The code that checked for SerialEvent being defined
in a sketch was only in the begin method where format was not passed in.

Changed these to have only one begin method, with an optional format descriptor
If format is not specified I pass in: (uint32_t)-1 which the begin method checks
for if it is -1 format is not called otherwise it is.

Note: with T4.x the default was always 0 and the format code was always done, but did not want to this here on the off chance there are T3.x sketches out there where
maybe the first call SerialX.begin(baud, some-not-default-format)... and later call
SerialX.begin(baud), and expected the format not to change.