Add Mix&Match option for USB
This patch allows to freely select keyboard (with or without
multimedia keys), mouse, joystick, serial port, and debug, in a
mix&match way, allowing to have any combination of those USB
interfaces (as long as your microcontroller has enough endpoints,
of course).

It uses some C PreProcessor magic to dinamically assign the
endpoints, addresses and everything else, as it can be seen
in usb_private.h, lines 116 onward, and usb.c, lines 288 onward.

To use this mode, select "USB Mix and Match (M&M)" as the USB
type, and then enable or disable the M&M options, according to
your needs.

Currently it's only implemented for teensy 2.0, but should be
straightforward to implement for all others, and also to add
all the other interfaces (disc, MIDI...).