Add fgets to File Impl in FS.h
Ran across an issue that fgets was not supported.  Looking through the forum seems like a few other people ok maybe one or two.  But it also looks like a handy function to have available so added it.  You can look here for a quick test:
https://forum.pjrc.com/index.php?threads/problems-with-external-library-using-sdfat.77492/post-363551
Will make a similar change to the SD library.